backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["aftern0on <faa31.31.31.31@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
sqlparse = "^0.4.2"
djangorestframework = "^3.12.4"
python-dotenv = "^0.19.1"
psycopg2 = "^2.9.1"
Pillow = "^8.4.0"
django-tinymce = "^3.3.0"
drf-spectacular = "^0.20.2"
django-filter = "^21.1"
transliterate = "^1.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

